# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuraAI Orchestrator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
  tool: 'tool',
} as const;

export interface Message {
  id: string;
  role: MessageRole;
  content: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  tokensUsed?: number | null;
  /** @nullable */
  latencyMs?: number | null;
  createdAt: string;
}

export type ChatInputMode = typeof ChatInputMode[keyof typeof ChatInputMode];


export const ChatInputMode = {
  chat: 'chat',
  code: 'code',
  vision: 'vision',
} as const;

export interface ChatInput {
  /** @nullable */
  sessionId?: string | null;
  message: string;
  /** @nullable */
  model?: string | null;
  mode: ChatInputMode;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  temperature?: number | null;
  /** @nullable */
  maxTokens?: number | null;
  ragEnabled: boolean;
  toolsEnabled: boolean;
  tools?: string[];
}

export interface ChatCompletion {
  message: Message;
  sessionId: string;
  model: string;
  ragChunksUsed: number;
  toolCallsExecuted: number;
}

export type ModelType = typeof ModelType[keyof typeof ModelType];


export const ModelType = {
  chat: 'chat',
  code: 'code',
  vision: 'vision',
} as const;

export interface Model {
  id: string;
  displayName: string;
  type: ModelType;
  available: boolean;
  defaultTemperature: number;
  defaultMaxTokens: number;
  description: string;
}

export interface EmbeddingInput {
  text: string;
  /** @nullable */
  model?: string | null;
}

export interface EmbeddingResponse {
  embedding: number[];
  model: string;
  dimensions: number;
}

export interface DocumentInput {
  title: string;
  content: string;
  /** @nullable */
  source?: string | null;
  tags?: string[];
}

export interface Document {
  id: string;
  title: string;
  content: string;
  /** @nullable */
  source?: string | null;
  tags?: string[];
  chunkCount: number;
  createdAt: string;
}

export interface RagQueryInput {
  query: string;
  topK?: number;
}

export interface RagChunk {
  documentId: string;
  documentTitle: string;
  content: string;
  score: number;
}

export interface RagResult {
  chunks: RagChunk[];
  query: string;
}

export type ToolType = typeof ToolType[keyof typeof ToolType];


export const ToolType = {
  http: 'http',
  shell: 'shell',
  git: 'git',
} as const;

export type ToolInputSchema = { [key: string]: unknown };

export interface Tool {
  name: string;
  description: string;
  type: ToolType;
  inputSchema: ToolInputSchema;
}

export type ToolRunInputInput = { [key: string]: unknown };

export interface ToolRunInput {
  tool: string;
  input: ToolRunInputInput;
}

export interface ToolRunResponse {
  tool: string;
  success: boolean;
  output: string;
  /** @nullable */
  error?: string | null;
  durationMs: number;
}

export type SessionMode = typeof SessionMode[keyof typeof SessionMode];


export const SessionMode = {
  chat: 'chat',
  code: 'code',
  vision: 'vision',
} as const;

export interface Session {
  id: string;
  title: string;
  model: string;
  mode: SessionMode;
  /** @nullable */
  systemPrompt?: string | null;
  temperature: number;
  maxTokens: number;
  ragEnabled: boolean;
  toolsEnabled: boolean;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export type SessionInputMode = typeof SessionInputMode[keyof typeof SessionInputMode];


export const SessionInputMode = {
  chat: 'chat',
  code: 'code',
  vision: 'vision',
} as const;

export interface SessionInput {
  title: string;
  /** @nullable */
  model?: string | null;
  mode?: SessionInputMode;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  temperature?: number | null;
  /** @nullable */
  maxTokens?: number | null;
  ragEnabled?: boolean;
  toolsEnabled?: boolean;
}

export interface SessionUpdate {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  mode?: string | null;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  temperature?: number | null;
  /** @nullable */
  maxTokens?: number | null;
  /** @nullable */
  ragEnabled?: boolean | null;
  /** @nullable */
  toolsEnabled?: boolean | null;
}

export interface SessionDetail {
  session: Session;
  messages: Message[];
}

export interface ModelUsageStat {
  model: string;
  count: number;
  tokens: number;
}

export interface ActivityItem {
  type: string;
  description: string;
  createdAt: string;
}

export interface AdminStats {
  totalSessions: number;
  totalMessages: number;
  totalDocuments: number;
  totalToolRuns: number;
  avgResponseMs: number;
  modelUsage: ModelUsageStat[];
  recentActivity: ActivityItem[];
}

export type LogEntryType = typeof LogEntryType[keyof typeof LogEntryType];


export const LogEntryType = {
  chat: 'chat',
  tool: 'tool',
  rag: 'rag',
  error: 'error',
} as const;

export interface LogEntry {
  id: string;
  type: LogEntryType;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  model?: string | null;
  summary: string;
  /** @nullable */
  latencyMs?: number | null;
  /** @nullable */
  tokensUsed?: number | null;
  success: boolean;
  createdAt: string;
}

export interface OllamaStatus {
  connected: boolean;
  baseUrl: string;
  availableModels: string[];
  /** @nullable */
  version?: string | null;
  /** @nullable */
  error?: string | null;
}