# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AuraAI Orchestrator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/rag/ingest:
    post:
      operationId: ingestDocument
      tags: [rag]
      summary: Ingest a document for RAG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
...
  /v1/rag/documents/{documentId}:
    delete:
      operationId: deleteDocument
      tags: [rag]
      summary: Delete an ingested document
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Document deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/rag/query:
    post:
      operationId: ragQuery
      tags: [rag]
      summary: Query the RAG knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RagQueryInput"
      responses:
        "200":
          description: Retrieved chunks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagResult"
...
  /v1/tools/run:
    post:
      operationId: runTool
      tags: [tools]
      summary: Execute a tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolRunInput"
      responses:
        "200":
          description: Tool execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolRunResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    DocumentInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        source:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - content
    Document:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        source:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        chunkCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - content
        - chunkCount
        - createdAt
    RagQueryInput:
      type: object
      properties:
        query:
          type: string
        topK:
          type: integer
      required:
        - query
    RagChunk:
      type: object
      properties:
        documentId:
          type: string
        documentTitle:
          type: string
        content:
          type: string
        score:
          type: number
      required:
        - documentId
        - documentTitle
        - content
        - score
    RagResult:
      type: object
      properties:
        chunks:
          type: array
          items:
            $ref: "#/components/schemas/RagChunk"
        query:
          type: string
      required:
        - chunks
        - query
...
    ToolRunInput:
      type: object
      properties:
        tool:
          type: string
        input:
          type: object
          additionalProperties: true
      required:
        - tool
        - input
    ToolRunResponse:
      type: object
      properties:
        tool:
          type: string
        success:
          type: boolean
        output:
          type: string
        error:
          type: ["string", "null"]
        durationMs:
          type: integer
      required:
        - tool
        - success
        - output
        - durationMs
...