# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuraAI Orchestrator API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: string;
  title: string;
  content: string;
  /** @nullable */
  source?: string | null;
  tags?: string[];
  chunkCount: number;
  createdAt: string;
}