# lib/api-zod/src/generated/types/ragQueryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuraAI Orchestrator API
 * OpenAPI spec version: 0.1.0
 */

export interface RagQueryInput {
  query: string;
  topK?: number;
}