# lib/api-zod/src/generated/types/ragResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuraAI Orchestrator API
 * OpenAPI spec version: 0.1.0
 */
import type { RagChunk } from './ragChunk';

export interface RagResult {
  chunks: RagChunk[];
  query: string;
}