# lib/api-zod/src/generated/types/toolRunInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuraAI Orchestrator API
 * OpenAPI spec version: 0.1.0
 */
import type { ToolRunInputInput } from './toolRunInputInput';

export interface ToolRunInput {
  tool: string;
  input: ToolRunInputInput;
}